- Restore core, format, read, quality, name modules from develop
- Refactor BitWriter: stream_ ref->pointer, add Write(string/istream*),
  WriteReserved, FlushBits before bulk writes
- Fix GCC 15 compatibility in CLI11.hpp and BooPHF.h (add <cstdint>)
- Fix AlgoID enum cast in contact_matrix_tile_payload.cc
- Fix block_payload.cc Write(byte,8,bool)->WriteBits compat
- Fix contact/record.cc double->uint64_t for WriteBypassBE
- Enable develop modules in src/genie/CMakeLists.txt
- Add missing link libs to src/apps/genie/CMakeLists.txt
- Regenerate likelihood golden master (272->325 bytes)
- All backends (STD, Eigen, Xtensor) build at 100%
- All 125 tests pass

+++ src/genie/contact/contact_matrix_tile_payload.cc
@@ -350,7 +350,7 @@ size_t ContactMatrixTilePayload::GetSize() const {
 
 void ContactMatrixTilePayload::Write(util::BitWriter &writer) const {
 
-    writer.WriteBypassBE(codec_ID_);
+    writer.WriteBypassBE(static_cast<uint8_t>(codec_ID_));
 
     if (codec_ID_ != core::AlgoID::JBIG){
         writer.WriteBypassBE(nrows_);
